Signals, Denormalization, and Logging
Moved away from saving the updates in the save method of the session model in favour of going with some pre and save signals.

Apparently this decentralized way of keeping values updated as each session object updates is Denormalization which is different to the more obvious way of just re-tallying all the session values in a central place.